==> 00_stock_data.ipynb <==
import pandas as pd
import os
from sklearn.model_selection import train_test_split
import numpy as np

os.chdir('../../Data/Nifty')

development = pd.DataFrame()
for file in sorted(os.listdir()):
    df = pd.read_csv(file,header=0)
    df = df[df['Series']=='EQ']
    df['Date'] = pd.to_datetime(df['Date'])
    df.sort_values(by='Date',inplace=True)
    development = development.append(df)

development['lift'] = (development['Close Price'] - development['Open Price']).apply(lambda x: 1 if x>0 else 0)

os.chdir('..')

os.getcwd()

development.to_csv('stock_data',index=False)
