==> 01_train_test_split.ipynb <==
import pandas as pd
import os
from sklearn.model_selection import train_test_split
import numpy as np

os.chdir('../../Data')

df = pd.read_csv('final_data_oot.csv',header=0)

df.columns

rename_dict = {col : '_'.join(col.split(' ')) for col in df.columns}

df.rename(rename_dict,copy=False,inplace=True,axis=1)

companies = list(df['Symbol'].unique())

cols_to_shift = [col for col in df.columns if col not in ['Date','Symbol','Series','Prev_Close','Open','Close','Sector']]

rename_dict2 = {col : col+'_pct_change' for col in cols_to_shift}

rename_dict2

development = pd.DataFrame()
for company in companies:
    print('development data for {}'.format(company))
    df_append = df[df['Symbol']==company]
    df_append['Date'] = pd.to_datetime(df['Date'])
    df_append.sort_values(by='Date',inplace=True)
    df_append['Prev_Close_max'] = df_append['Prev_Close'].rolling(5).max()
    df_append['Prev_Close_min'] = df_append['Prev_Close'].rolling(5).min()
    df_append['Prev_Close_max_ratio'] = df_append['Prev_Close']/df_append['Prev_Close_max']
    df_append['Prev_Close_min_ratio'] = df_append['Prev_Close_min']/df_append['Prev_Close']
    for index in ['nifty','auto','bank','fmcg','it','media','metal','pharma','realty','psu','500']:
        for i in range(1,6):
            df_append['return_{j}d_{index}'.format(j=i,index=index)] = (df_append['Close_{}'.format(index)].shift(i)-df_append['Open_{}'.format(index)].shift(i))/df_append['Open_{}'.format(index)].shift(i)
        df_append['Voltality_{}'.format(index)] = df_append[['return_{i}d_{index}'.format(i=i,index=index) for i in range(1,6)]].apply(lambda x:np.std(x),axis=1)
    for i in range(1,6):
        df_append['return_{j}d'.format(j=i)] = (df_append['Close'].shift(i)-df_append['Open'].shift(i))/df_append['Open'].shift(i)
    df_append['Voltality'] = df_append[['return_1d','return_2d','return_3d','return_4d','return_5d']].apply(lambda x:np.std(x),axis=1)
    df_append['Open_pct_change'] = df_append['Open'].pct_change().shift(1)
    df_append['Prev_Close_pct_change'] = df_append['Prev_Close'].pct_change()
    for col in cols_to_shift:
        df_append[col] = df_append[col].pct_change().shift(1)
    df_append.rename(rename_dict2,copy=False,inplace=True,axis=1)
    development = development.append(df_append.iloc[5:])
development['actual_return'] = (development['Close']-development['Open'])/development['Open']

development.isnull().any()

development.shape

development.columns

null_cols = []
for col in development.columns:
    if development[col].isnull().any():
        print(col)
        null_cols.append(col)

null_cols

development.drop(null_cols,axis=1,inplace=True)

for col in development.columns:
    if development[col].isnull().any():
        print(col)

training = development.loc[development['Date']< '2019-01-01']
testing =  development.loc[development['Date']>='2019-01-01']

os.chdir('../')

training.shape

os.chdir('Price Prediction/Input_Data/')

development.to_csv('oot.csv',index=False)

training.to_csv('train.csv',index=False)

testing.to_csv('validation.csv',index=False)

development.head()

for col in [col for col in development.columns if col not in ['Date','Symbol','Series','Prev_Close','Open','Close','Sector']]:
    print("{col} : min = {min}, max = {max}".format(col=col,min=development[col].min(),max=development[col].max()))
