==> 03_outlier_treatment_imputation.ipynb <==
import pandas as pd
import numpy as np
import pickle
import os
from scipy.stats import skew,kurtosis
import matplotlib.pyplot as plt
from sklearn.linear_model import LinearRegression
from advanced_analysis_package import variable_treatment,analyze

from openpyxl import load_workbook,Workbook

os.chdir('../output_data/')

df = pd.read_csv('train_v01.csv',header=0)

df.columns

os.chdir('../pickles')
numerical = pickle.load(open('numerical.pickle','rb'))
categorical = pickle.load(open('categorical.pickle','rb'))

os.chdir('../Statistics')

ignore_cols = ['Symbol','Series','Date','Prev_Close','Open','Close','Prev_Close_min','Prev_Close_max']

def automatic_outlier_treatment(data,s):
    
    lower_limit = data[s].quantile(q=.25)-1.5*(data[s].quantile(q=.75)-data[s].quantile(q=.25))
    upper_limit = data[s].quantile(q=.75)+1.5*(data[s].quantile(q=.75)-data[s].quantile(q=.25))

    frac_lower = float(np.unique(np.array(data.loc[(data[s].notnull()) & (data[s] < lower_limit),s]).shape[0]))/data.shape[0]
    frac_upper = float(np.unique(np.array(data.loc[(data[s].notnull()) & (data[s] > upper_limit),s]).shape[0]))/data.shape[0]

    p1,low_sigma3,low_sigma2 = data[s].quantile(q=.01),data[s].mean()-3*data[s].std(),data[s].mean()-2*data[s].std()
    p99,high_sigma3,high_sigma2 = data[s].quantile(q=.99),data[s].mean()+3*data[s].std(),data[s].mean()+2*data[s].std()
    p5,p95 = data[s].quantile(q=.05),data[s].quantile(q=.95)
    minima,maxima = data[s].min(),data[s].max()

    #deriving lower limit in case of aggresive treatment
    if low_sigma3 >= minima:
        b = min(p1,low_sigma3)
    elif low_sigma2>= minima:
        b = min(p1,low_sigma2)
    else:
        b=p1

    #deriving upper limit in case of aggresive treatment
    if high_sigma3 <= maxima:
        a = max(p99,high_sigma3)
    elif high_sigma2<= maxima:
        a = max(p99,high_sigma2)
    else:
        a=p99

    outlier_treatment = ''

    #lower end outlier treatment
    if frac_lower<=.05:
        variable_treatment.capping_and_flooring(data,s,a,b,high=False,low=True)
        if b==p1:
            outlier_treatment+='flooring at p1,'
        elif b==low_sigma2:
            outlier_treatment+='flooring at sigma2,'
        elif b==low_sigma3:
            outlier_treatment+='flooring at sigma3,'
    else:
        try:
            variable_treatment.exponential_smoothning(data,s,low=True,high=False)
            outlier_treatment+='lower line exponential smoothing,'
        except:
            pass

    #higher end outlier treatment
    if frac_upper<=.05:
        variable_treatment.capping_and_flooring(data,s,a,b,high=True,low=False)
        if b==p1:
            outlier_treatment+='capping at p99'
        elif b==low_sigma2:
            outlier_treatment+='capping at sigma2'
        elif b==low_sigma3:
            outlier_treatment+='Capping at sigma3'
    else:
        try:
            variable_treatment.exponential_smoothning(data,s,low=False,high=True)
            outlier_treatment+='higher line exponential smoothing'
        except:
            pass

    return outlier_treatment

numerical,categorical = analyze.numerical_categorical_division(df.drop(ignore_cols,axis=1))

categorical

numerical = [x for x in numerical if (x != 'actual_return' and x not in categorical)]

df[numerical].dtypes

edd_df = pd.read_excel('report.xlsx',sheet_name='edd_price_prediction',header=0)

edd_df['outlier_treatment'] = ''
for col in numerical:
    edd_df.loc[edd_df['Var']==col,'outlier_treatment']=automatic_outlier_treatment(df,col)

numerical

edd_df['outlier_treatment'].value_counts()

wb = load_workbook('report.xlsx')
writer = pd.ExcelWriter('report.xlsx',engine='openpyxl')
writer.book = wb
edd_df.to_excel(writer,sheet_name='edd_price_prediction',index=False)
writer.save()
writer.close()

os.chdir('../output_data')

df.to_csv('train_v02.csv',index=False)

df = pd.read_csv('train_v03.csv',header=0)

numerical,categorical = analyze.numerical_categorical_division(df.drop(ignore_cols,axis=1))

for col in numerical:
    df[col] = df[col].apply(lambda x : x if np.abs(x) <= 1 else np.sign(x)*1)

categorical

numerical

df.to_csv('train_v03.csv',index=False)

# Imputation

conversion_dict={}

numerical

categorical

for col in categorical:
    conversion_dict[col]=variable_treatment.make_dummies(df,col)

os.chdir('../pickles')

pickle.dump(conversion_dict,open('conversion_dict.pickle','wb'))
pickle.dump(numerical,open('numerical.pickle','wb'))
pickle.dump(categorical,open('categorical.pickle','wb'))

conversion_dict

imputation_cols = [col for col in df.columns if col not in numerical+['actual_return']+categorical]

imputation_cols

df

pickle.dump(imputation_cols,open('imputation_cols.pickle','wb'))

os.chdir('../Imputation_models')

for col in [x for x in numerical if x!='actual_return']:
    data = df[imputation_cols+[col]].dropna(subset=[col])
    x = np.array(data[imputation_cols])
    y = np.array(data[col])
    model = LinearRegression()
    model.fit(x,y)
    del x
    del y
    pickle.dump(model,open(col+'_impute.pickle','wb'))
    if df[col].isnull().any():
        indices = df.loc[df[col].isnull()].index.tolist()
        df.loc[indices,col]=np.array(model.predict(np.array(df.loc[indices,imputation_cols])))
        del indices
    del model
    print(col+' imputed')

df[[x for x in numerical if x!='actual_return']].isnull().any()

default_categorical

df.drop(default_categorical,axis=1,inplace=True)

os.chdir('../output_data')

rename_dict = {col : col.replace(' ','_').replace('-','_').replace('&','_') for col in df.columns if '-' in col or ' ' in col}

rename_dict

df.rename(rename_dict,axis=1,inplace=True)

df.columns

df.to_csv('train_v04.csv',index=False)

df.isnull().any()
