==> 06_model_evaluation.ipynb <==
import pandas as pd
import numpy as np
import pickle
import os

from sklearn.metrics import r2_score,mean_squared_error,precision_score,recall_score,f1_score,accuracy_score

from scipy.special import cbrt

os.chdir('../pickles')

numerical = pickle.load(open('numerical.pickle','rb'))
categorical = pickle.load(open('categorical.pickle','rb'))
conversion_dict = pickle.load(open('conversion_dict.pickle','rb'))
imputation_cols = pickle.load(open('imputation_cols.pickle','rb'))
model_columns = pickle.load(open('model_columns.pickle','rb'))
lr_model_columns = pickle.load(open('lr_model_columns.pickle','rb'))
order = pickle.load(open('order.pickle','rb'))
pca = pickle.load(open('pca.pickle','rb'))
scale = pickle.load(open('scale.pickle','rb'))

os.chdir('../Input_Data')

df = pd.read_csv('validation.csv',header=0)

rename_dict = {col : col.replace('%','pct_') for col in df.columns if '%' in col}

df_val = df

df_val.rename(rename_dict,axis=1,inplace=True)

df_val['lift'] = df_val['actual_return'].apply(lambda x: 1 if x >0 else 0) 

for col in categorical:
    for category in conversion_dict[col]:
        df_val[col+'_dum_'+str(category)] = 0
        df_val.loc[df_val[col]==category,col+'_dum_'+str(category)]=1
    print(col+' done')

rename_dict = {col : col.replace(' ','_').replace('-','_').replace('&','_') for col in df.columns if '-' in col or ' ' in col}

df_val.rename(rename_dict,axis=1,inplace=True)

for col in model_columns:
    df_val[col] = df_val[col].apply(lambda x : x if np.abs(x) <= 1 else np.sign(x)*1)

x,y = np.array(df_val[model_columns]),np.array(df_val['actual_return'])

os.chdir('../Models/')

model = pickle.load(open('LR_wr.pickle','rb'))

y_pred = model.predict(x)

r2_score(y,y_pred)

df_val['lift_predicted'] = [max(0,np.sign(x)) for x in y_pred]

accuracy_score(y_true=df_val['lift'],y_pred=df_val['lift_predicted'])

precision_score(y_true=df_val['lift'],y_pred=df_val['lift_predicted'])

recall_score(y_true=df_val['lift'],y_pred=df_val['lift_predicted'])

df_val['month'] = df_val['Date'].apply(lambda x : str(x)[5:7])
df_val['year'] = df_val['Date'].apply(lambda x: str(x)[0:4])

df_val['correct_prediction'] = df_val[['lift','lift_predicted']].apply(lambda x: 1 if int(x[0]) == int(x[1]) else 0,axis=1)

df_val['precise_prediction'] = df_val[['lift','lift_predicted']].apply(lambda x: 1 if int(x[0]) == 1 and int(x[1]) == 1 else 0,axis=1)

df_lr_grp = df_val.groupby(['month','year','Symbol','Sector']).agg({'lift':['count','sum'],'lift_predicted':['sum'],
                                                          'correct_prediction':['sum'],'precise_prediction':['sum']})

df_lr_grp.reset_index(inplace=True)

df_lr_grp.columns = ['month','year','Symbol','Sector','trade_days','price_lift','pred_price_lift','correct_predcitions','precise_predictions']

df_lr_grp

os.chdir('../Statistics/')

from openpyxl import Workbook,load_workbook

wb = load_workbook('report.xlsx')
writer = pd.ExcelWriter('report.xlsx',engine='openpyxl')
writer.book = wb
df_lr_grp.to_excel(writer,sheet_name='lr_price_stats_val_wr',index=False)
writer.save()
writer.close()

# for col in categorical:
#     if len(conversion_dict[col])==1:
#         category = conversion_dict[col][0]
#         df_val[col+'_dum_'+str(category)] = 0
#         df_val.loc[df_val[col]==category,col+'_dum_'+str(category)]=1
#     else:
#         total_categories = len(conversion_dict[col])
#         dummies = len(str(int(bin(total_categories)[2:],10)))
#         bin_conv=[]
#         for i in range(total_categories):
#             bin_conv.append(conversion_dict[col][i][1])
#         for j in range(dummies):
#             df_val[col+'_dum_'+str(j)]=0
#             for i,cat in enumerate([conv[0] for conv in conversion_dict[col]]):
#                 df_val.loc[df_val[col]==cat,col+'_dum_'+str(j)]=bin_conv[i]%10
#                 bin_conv[i]=bin_conv[i]//10
#     df_val.drop(col,axis=1,inplace=True)
#     print(col+' done')

os.chdir('../Statistics/')

edd_df = pd.read_excel('report.xlsx',sheet_name='edd_v01',header=0)

transform_dict = {'log':lambda x: np.log(x),'sqr':lambda x: x**2,'sqrt':lambda x: np.sqrt(x),'exp':lambda x:np.exp(x),
                 'cube':lambda x: x**3,'cuberoot': lambda x: cbrt(x)}

for col in list(edd_df['Var']):
    applied = edd_df.loc[edd_df['Var']==col,'conversions'].values[0]
    if applied != '' and applied != 'categorical':
        try:
            for t in applied.split(','):
                df_val[col+'_'+t] = df_val[col].apply(transform_dict[t])
        except:
            pass

os.chdir('../Imputation_models')

for col in [x for x in numerical if x!='actual_return']:
    model = pickle.load(open(col+'_impute.pickle','rb'))
    if df_val[col].isnull().any():
        indices = df_val.loc[df_val[col].isnull()].index.tolist()
        df_val.loc[indices,col]=np.array(model.predict(np.array(df_val.loc[indices,imputation_cols])))
        del indices
    del model
    print(col+' imputed')

x_val,y_val = np.array(df_val[model_columns]),np.array(df_val['actual_return'])
x,y = np.array(df_val[lr_model_columns]),np.array(df_val['actual_return'])
x_pca,y_pca = pca.transform(scale.transform(df_val[model_columns])),np.array(df_val['actual_return'])

os.chdir('../Models')

models = {'LR':pickle.load(open('LR.pickle','rb')),
                      'RF':pickle.load(open('RF.pickle','rb')),
                    'GBM':pickle.load(open('GBM.pickle','rb')),
          'RF_full':pickle.load(open('RF_full.pickle','rb')),
          'GBM_full':pickle.load(open('GBM_full.pickle','rb')),
          'RF_pca':pickle.load(open('RF_pca.pickle','rb')),
          'GBM_pca':pickle.load(open('GBM_pca.pickle','rb'))
         }

df_val['RF_Prediction'] = models['RF'].predict(x)

df_val['GBM_Prediction'] = models['GBM'].predict(x)

df_val['LR_Prediction'] = models['LR'].predict(x)

df_val['RF_full_Prediction'] = models['RF_full'].predict(x_val)

df_val['GBM_full_Prediction'] = models['GBM_full'].predict(x_val)

df_val['RF_pca_Prediction'] = models['RF_pca'].predict(x_pca)

df_val['GBM_pca_Prediction'] = models['GBM_pca'].predict(x_pca)

scores = [r2_score(df_val['actual_return'],df_val['LR_Prediction']),r2_score(df_val['actual_return'],df_val['RF_Prediction']),r2_score(df_val['actual_return'],df_val['GBM_Prediction']),
         r2_score(df_val['actual_return'],df_val['RF_full_Prediction']),r2_score(df_val['actual_return'],df_val['GBM_full_Prediction']),
          r2_score(df_val['actual_return'],df_val['RF_pca_Prediction']),r2_score(df_val['actual_return'],df_val['GBM_pca_Prediction'])
         ]

df_scores = pd.DataFrame(list(zip(['LR','RF','GBM','RF_full','GBM_full','RF_pca','GBM_pca'],scores)),columns=['models','r2_score'])

df_scores

os.chdir('../Input_Data/')

open_prices = pd.read_csv('open_price.csv',header=None)

open_prices.columns = ['Symbol','Date','Open','Close']

df_val[['Symbol','Date']] = df[['Symbol','Date']]

df_final = pd.merge(df_val,open_prices,on=['Symbol','Date'])

def loss_precision(model):
    expected_profits = df_final[model+'_Prediction']
    actual_profits = df_final['actual_return']
    expected_lift = expected_profits.apply(lambda x: 1 if x >0 else 0)
    actual_lift = actual_profits.apply(lambda x: 1 if x>0 else 0)
    precision = precision_score(y_true=actual_lift,y_pred=expected_lift)
    recall = recall_score(y_true=actual_lift,y_pred=expected_lift)
    f1 = f1_score(y_true=actual_lift,y_pred=expected_lift)
    return [precision,recall,f1]

df_scores['precision'] = df_scores['models'].apply(lambda x: loss_precision(x)[0])
df_scores['recall']=df_scores['models'].apply(lambda x: loss_precision(x)[1])
df_scores['f1_score']=df_scores['models'].apply(lambda x: loss_precision(x)[2])

df_scores

df_final['ensemble_prediction'] = (df_final['LR_Prediction']*.45+df_final['RF_Prediction']*.61+df_final['GBM_Prediction']*.49
                                  +df_final['RF_full_Prediction']*0+df_final['GBM_full_Prediction']*.47+
                                  df_final['RF_pca_Prediction']*.53+df_final['GBM_pca_Prediction']*.40)/(.45+.61+.49+.47+.53+.4)

expected_profits = df_final['ensemble_prediction']
actual_profits = df_final['actual_return']
expected_lift = expected_profits.apply(lambda x: 1 if x >0 else 0)
actual_lift = actual_profits.apply(lambda x: 1 if x>0 else 0)
precision = precision_score(y_true=actual_lift,y_pred=expected_lift)
recall = recall_score(y_true=actual_lift,y_pred=expected_lift)
f1 = f1_score(y_true=actual_lift,y_pred=expected_lift)

precision

df_scores

append_list = ['Ensembled',r2_score(df_final['Close'],df_final['ensemble_prediction']),precision,recall,f1]

df_append = pd.DataFrame([append_list],columns=df_scores.columns)

df_scores = df_scores.append(df_append,ignore_index=True)

df_scores

from openpyxl import load_workbook,Workbook

os.chdir('../Statistics/')

rf_importance = pd.DataFrame(list(zip(lr_model_columns,list(models['RF'].feature_importances_))),columns=['Var','Importance'])
book = load_workbook('report.xlsx')
writer = pd.ExcelWriter('report.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

rf_importance.to_excel(writer, "rf_feature_importance", index=False)

writer.save()

gbm_importance = pd.DataFrame(list(zip(lr_model_columns,list(models['GBM'].feature_importances_))),columns=['Var','Importance'])
book = load_workbook('report.xlsx')
writer = pd.ExcelWriter('report.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

gbm_importance.to_excel(writer, "gbm_feature_importance", index=False)

writer.save()

linear_importance = pd.DataFrame(list(zip([1]+list(lr_model_columns),[models['LR'].intercept_]+list(models['LR'].coef_))),columns=['Var','coeff'])
book = load_workbook('report.xlsx')
writer = pd.ExcelWriter('report.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

linear_importance.to_excel(writer, "linear_model_coef", index=False)

writer.save()

rf_importance = pd.DataFrame(list(zip(model_columns,list(models['RF_full'].feature_importances_))),columns=['Var','Importance'])
book = load_workbook('report.xlsx')
writer = pd.ExcelWriter('report.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

rf_importance.to_excel(writer, "rf_full_feature_importance", index=False)

writer.save()

gbm_importance = pd.DataFrame(list(zip(model_columns,list(models['GBM_full'].feature_importances_))),columns=['Var','Importance'])
book = load_workbook('report.xlsx')
writer = pd.ExcelWriter('report.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

gbm_importance.to_excel(writer, "gbm_full_feature_importance", index=False)

writer.save()

book = load_workbook('report.xlsx')
writer = pd.ExcelWriter('report.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

df_scores.to_excel(writer, "model_scores", index=False)

writer.save()

os.chdir('../output_data/')

df_final.columns

df_final.to_csv('validation_predicted.csv',index=False)

os.chdir('../Statistics/')

df_final.to_excel('Q1_predictions.xlsx',sheet_name='data',index=False)
