==> 07_reason_codes.ipynb <==
import pandas as pd
import numpy as np
import os
import pickle
from treeinterpreter import treeinterpreter as ti
import shap
from scipy.special import cbrt

from openpyxl import load_workbook,Workbook

# Getting data

os.chdir('../pickles')

numerical = pickle.load(open('numerical.pickle','rb'))
categorical = pickle.load(open('categorical.pickle','rb'))
conversion_dict = pickle.load(open('conversion_dict.pickle','rb'))
imputation_cols = pickle.load(open('imputation_cols.pickle','rb'))
model_columns = pickle.load(open('model_columns','rb'))
lr_model_columns = pickle.load(open('lr_model_columns.pickle','rb'))
order = pickle.load(open('order.pickle','rb'))

os.chdir('../Statistics/')

df_val = pd.read_excel('Q1_predictions.xlsx',sheet_name='data',header=0)

df_val.columns

os.chdir('../Input_Data/')

# Reason Codes for Random Forest Model

os.chdir('../Models/')

RF = pickle.load(open('RF.pickle','rb'))

df_val.reset_index(inplace=True)

df_val.columns

def reason_code_rf(model,data,target='Close'):
    predictors = lr_model_columns
    df_reasons = pd.DataFrame()
    prediction, bias, contributions = ti.predict(model,np.array(data[lr_model_columns]))
    df_reasons['Symbol'] = data['Symbol']
    df_reasons['Date'] = data['Date']
    df_reasons['Open Price'] = data['Open']
    df_reasons['Prev Close'] = data['Prev_Close(in 10^5)']*(10**5)
    df_reasons['bias'] = bias
    df_reasons['Actual_Value'] = data['Close']
    df_reasons['Predicted_Value']=prediction
    df_reasons['Expected Profit'] = df_reasons['Predicted_Value']-df_reasons['Prev Close']
    df_reasons['Actual Profit'] = df_reasons['Actual_Value'] - df_reasons['Open Price']
    df_reasons['Profit Diff'] = df_reasons['Expected Profit'] - df_reasons['Actual Profit']
    contributions = pd.DataFrame(contributions)
    contributions['mrc'] = contributions.apply(lambda x: ','.join([a+':'+str(b) for (a,b) in sorted(zip(lr_model_columns,x),key = lambda x: x[1])]),axis=1)
    df_reasons['mrc'] = contributions['mrc']
    
    return df_reasons

df_reasons = reason_code_rf(RF,df_val)

df_reasons

os.chdir('../Statistics/')

book = load_workbook('Q1_predictions.xlsx')
writer = pd.ExcelWriter('Q1_predictions.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

df_reasons.to_excel(writer, "rf_reason_codes", index=False)

writer.save()

# Reason Codes for Gradient Boosting Model

os.chdir('../Models/')

GBM = pickle.load(open('GBM.pickle','rb'))

os.chdir('../output_data/')

def reason_code_gbm(model,data,target='Close_Price'):
    predictors = lr_model_columns
    df_reasons = pd.DataFrame()
    df_reasons['Symbol'] = data['Symbol']
    df_reasons['Date'] = data['Date']
    df_reasons['Open Price'] = data['Open']
    df_reasons['Prev Close'] = data['Prev_Close(in 10^5)']*(10**5)
    df_reasons['Actual_Value'] = data['Close']
    df_reasons['Predicted_Value']=model.predict(np.array(data[lr_model_columns]))
    df_reasons['Expected Profit'] = df_reasons['Predicted_Value']-df_reasons['Prev Close']
    df_reasons['Actual Profit'] = df_reasons['Actual_Value'] - df_reasons['Open Price']
    df_reasons['Profit Diff'] = df_reasons['Expected Profit'] - df_reasons['Actual Profit']
    explainer = shap.TreeExplainer(model)
    shap_values = explainer.shap_values(data[model_columns])
    df_reasons['bias'] = explainer.expected_value
    contributions = pd.DataFrame(shap_values)
    contributions['mrc'] = contributions.apply(lambda x: ','.join([a+':'+str(b) for (a,b) in sorted(zip(lr_model_columns,x),key = lambda x: x[1])]),axis=1)
    df_reasons['mrc'] = contributions['mrc']
    
    return df_reasons

df_reason = reason_code_gbm(GBM,df_val)

df_reason.head()

os.chdir('../Statistics/')

book = load_workbook('Q1_predictions.xlsx')
writer = pd.ExcelWriter('Q1_predictions.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

df_reason.to_excel(writer, "gbm_reason_codes", index=False)

writer.save()

# Reason Codes for linear regression model

os.chdir('../Models/')

LR = pickle.load(open('LR.pickle','rb'))

def reason_code_lr(model,data,target='Close'):
    df_reasons = pd.DataFrame()
    df_reasons['Symbol'] = data['Symbol']
    df_reasons['Date'] = data['Date']
    df_reasons['Open Price'] = data['Open']
    df_reasons['Prev Close'] = data['Prev_Close(in 10^5)']*(10**5)
    df_reasons['Actual_Value'] = data['Close']
    df_reasons['Predicted_Value'] = model.predict(data[lr_model_columns])
    df_reasons['Expected Profit'] = df_reasons['Predicted_Value']-df_reasons['Prev Close']
    df_reasons['Actual Profit'] = df_reasons['Actual_Value'] - df_reasons['Open Price']
    df_reasons['Profit Diff'] = df_reasons['Expected Profit'] - df_reasons['Actual Profit']
    df_reasons['bias'] = model.intercept_
    coef_array = np.array(model.coef_)
    values = np.array(data[lr_model_columns])
    contributions = np.array([coef_array*values[i,:] for i in range(values.shape[0])])
    contributions = pd.DataFrame(contributions)
    contributions['mrc'] = contributions.apply(lambda x: ','.join([a+':'+str(b) for (a,b) in sorted(zip(lr_model_columns,x),key = lambda x: x[1])]),axis=1)
    df_reasons['mrc'] = contributions['mrc']
    
    return df_reasons

df_reasons = reason_code_lr(LR,df_val)

df_reasons

os.chdir('../Statistics/')

book = load_workbook('Q1_predictions.xlsx')
writer = pd.ExcelWriter('Q1_predictions.xlsx', engine='openpyxl') 
writer.book = book
writer.sheets = dict((ws.title, ws) for ws in book.worksheets)

df_reasons.to_excel(writer, "lr_reason_codes", index=False)

writer.save()
